oca-port: moved to a package + split code in modules
